+++ R/7-aldex2propr.R
@@ -31,15 +31,16 @@
 #'  could otherwise change the proportionality measure.
 #'
 #' @param aldex.clr An \code{aldex.clr} object.
-#' @param how A character string. The proportionality method
-#'  used to build the \code{propr} object. For example,
-#'  "perb" returns an estimation of rho while "phit" returns
-#'  an estimation of phi.
-#' @param select See \code{\link{proportionality}}.
+#' @param how A character string. The proportionality metric
+#'  used to build the \code{propr} object. Choose from
+#'  "rho", "phi", or "phs".
+#' @param select Optional. Use this to subset the final
+#'  proportionality matrix without altering the result.
+#'
 #' @return Returns a \code{propr} object.
 #'
 #' @export
-aldex2propr <- function(aldex.clr, how = "perb", select){
+aldex2propr <- function(aldex.clr, how = "rho", select){
 
   packageCheck("ALDEx2")
 
@@ -101,14 +102,20 @@ aldex2propr <- function(aldex.clr, how = "perb", select){
   }
 
   propr <- new("propr")
-  propr@counts <- counts
-  propr@logratio <- logratio / k
+  propr@counts <- as.data.frame(counts)
+  propr@logratio <- as.data.frame(logratio) / k
   propr@matrix <- prop / k
+
+  message("Alert: Using 'aldex2propr' is not compatible the @results table.")
+  propr@results <- data.frame()
+
+  message("Alert: Using 'aldex2propr' disables permutation testing.")
+  propr@permutes <- list(NULL)
 
   return(propr)
 }
 
-#' Correlate CLR with a Continuous Measurement
+#' Correlate CLR Data with a Continuous Measurement
 #'
 #' This function correlates each log-ratio transformed
 #'  feature vector with a continuous numeric variable.
@@ -117,10 +124,10 @@ aldex2propr <- function(aldex.clr, how = "perb", select){
 #' @param conditions A numeric vector of a continuous variable.
 #' @param ... Arguments passed to \code{cor.test}.
 #'
-#' @return Returns a data.frame of the correlation
-#'  statistic (e.g., \code{r}) and p-value (\code{p})
-#'  for each log-ratio transformed feature.
-#'  FDR provided by \code{BH} column.
+#' @return Returns a data.frame of the estimated
+#'  correlation statistic (e.g., \code{r}) and p-value
+#'  (\code{p}) for each log-ratio transformed feature,
+#'  with FDR appended as the \code{BH} column.
 #'
 #' @importFrom stats cor.test p.adjust
 #' @export
@@ -148,7 +155,7 @@ lr2cor <- function(lr, conditions, ...){
              row.names = colnames(lr))
 }
 
-#' Correlate CLR with a Continuous Measurement
+#' Correlate CLR Data with a Continuous Measurement
 #'
 #' This function uses the Monte Carlo instances from an
 #'  \code{aldex.clr} object to correlate each log-ratio
@@ -160,10 +167,9 @@ lr2cor <- function(lr, conditions, ...){
 #' @param ... Arguments passed to \code{cor.test}.
 #'
 #' @return Returns a data.frame of the average
-#'  correlation statistic (e.g., \code{r}) and
-#'  average p-value (\code{p}) for each feature
-#'  across all Monte Carlo instances. Average
-#'  FDR provided by \code{BH} column.
+#'  correlation statistic (e.g., \code{r}) and p-value
+#'  (\code{p}) for each log-ratio transformed feature,
+#'  with FDR appended as the \code{BH} column.
 #'
 #' @export
 aldex.cor <- function(clr, conditions, ...){
@@ -181,7 +187,101 @@ aldex.cor <- function(clr, conditions, ...){
     r <- r + lr2cor(mci_lr, conditions, ...)
   }
 
-  r <- r / k
-  data.frame("r" = r$r, "p" = r$p, "BH" = r$BH,
-             row.names = colnames(mci_lr))
+  r / k # return expected
+}
+
+#' Build a Generalized Linear Model from CLR Data
+#'
+#' This function builds a generalized linear model from
+#'  log-ratio transformed data.
+#'
+#' @param lr A data.frame. A log-ratio transformed counts matrix.
+#' @param model.matrix A \code{model.matrix}.
+#' @param ... Arguments passed to \code{glm}.
+#'
+#' @return Returns a data.frame of the estimated
+#'  coefficients and their p-values for each feature,
+#'  with FDR appended as a \code{BH} column.
+#'
+#' @importFrom stats glm p.adjust coef
+#' @export
+lr2glm <- function(lr, model.matrix, ...){
+
+  if(class(model.matrix) != "matrix" &
+     !("assign" %in% names(attributes(model.matrix)))){
+
+    stop("Please use the 'model.matrix' function to prepare 'model.matrix'.")
+  }
+
+  if(nrow(lr) != nrow(model.matrix)){
+
+    stop("Input data and 'model.matrix' should have same number of rows.")
+  }
+
+  # Build the glm models
+  model. <- model.matrix
+  glms <- apply(lr, 2, function(x){
+    glm(x ~ model., ...)
+  })
+
+  # Extract coefficients and p-values
+  extract <- function(model){
+    x <- coef(summary(model))
+    coefs <- lapply(1:nrow(x), function(i){
+      y <- x[i,,drop=FALSE]
+      colnames(y) <- paste(rownames(y), colnames(y))
+      y})
+    do.call("cbind", coefs)
+  }
+
+  # Combine to make data.frame
+  extracts <- lapply(glms, extract)
+  df <- do.call("rbind", extracts)
+  rownames(df) <- colnames(lr)
+  df <- as.data.frame(df)
+
+  # Create new data.frame for FDR
+  pvals <- colnames(df)[grepl("Pr\\(>", colnames(df))]
+  df.bh <- df[,pvals]
+  colnames(df.bh) <- paste0(colnames(df.bh), ".BH")
+  for(j in 1:ncol(df.bh)){
+    df.bh[,j] <- p.adjust(df.bh[,j])
+  }
+
+  # Merge results with FDR
+  cbind(df, df.bh)
+}
+
+#' Build a Generalized Linear Model from CLR Data
+#'
+#' This function uses the Monte Carlo instances from an
+#'  \code{aldex.clr} object to build a generalized linear
+#'  model from log-ratio transformed data.
+#'  See \code{\link{lr2glm}}.
+#'
+#' @param clr An \code{aldex.clr} object.
+#' @param model.matrix A \code{model.matrix}.
+#' @param ... Arguments passed to \code{glm}.
+#'
+#' @return Returns a data.frame of the average
+#'  coefficients and their p-values for each feature,
+#'  with FDR appended as a \code{BH} column.
+#'
+#' @export
+aldex.glm <- function(clr, model.matrix, ...){
+
+  packageCheck("ALDEx2")
+
+  # Keep a running sum of lr2glm instances
+  mc <- ALDEx2::getMonteCarloInstances(clr)
+  k <- ALDEx2::numMCInstances(clr)
+  r <- 0
+  for(i in 1:k){
+
+    numTicks <- progress(i, k, numTicks)
+    mci_lr <- t(sapply(mc, function(x) x[, i]))
+    r <- r + lr2glm(mci_lr, model.matrix, ...)
+  }
+
+  r / k # return expected
 }