+++ inst/doc/a_introduction.R
@@ -4,21 +4,24 @@ N <- 100
 X <- data.frame(a=(1:N), b=(1:N) * rnorm(N, 10, 0.1),
                 c=(N:1), d=(N:1) * rnorm(N, 10, 1.0))
 
-## ---- dpi = 66, results = "hide"-----------------------------------------
+## ---- dpi = 66, results = "hide", message = FALSE------------------------
 library(propr)
-phi <- phit(X, symmetrize = TRUE)
-rho <- perb(X, ivar = 0)
-phs <- phis(X, ivar = 0)
+phi <- propr(X, metric = "phi", symmetrize = TRUE)
+rho <- propr(X, metric = "rho", ivar = 0)
+phs <- propr(X, metric = "phs", ivar = 0)
 
-## ---- dpi = 66-----------------------------------------------------------
+## ------------------------------------------------------------------------
+updateCutoffs(rho, cutoff =  seq(.05, .95, .3))
+
+## ---- dpi = 66, message = FALSE------------------------------------------
 rho99 <- rho[">", .95]
 rho99@pairs
 
-## ---- dpi = 66-----------------------------------------------------------
+## ---- dpi = 66, message = FALSE------------------------------------------
 rhoab <- subset(rho, select = c("a", "b"))
 rhoab@matrix
 
-## ---- dpi = 66-----------------------------------------------------------
+## ---- dpi = 66, message = FALSE------------------------------------------
 simplify(rho99)
 
 ## ---- dpi = 66, results = "hide", fig.show = "hold", fig.keep = "last"----