+++ NEWS.md
@@ -1,3 +1,41 @@
+## propr 4.0.0
+---------------------
+* Update `ivar2index` method
+    * Now replaces zeros with 1 to calculate IQR (fixes "iqlr" bug for alpha > 0)
+* Update `propr` object backend and API
+    * Heavily revise documentation to harmonize `propd` with `propr`
+    * Create `propr` function to replace `perb`, `phit`, and `phis`
+    * Add `@results` slot for storing proportionality half-matrix
+    * Add `@permutes` slot for storing reproducible permutations
+    * Update `subset` and `[` to disable `@results` and `@permutes`
+    * Add `@fdr` slot for storing FDR results
+    * Add `alpha` argument
+        * Adjusts alpha-based VLR by var[(component^alpha - reference^alpha)/alpha]
+        * Saves `alpha` to `@alpha` and alpha-based counts to `@logratio`
+        * Note this is not yet technically equivalent to `propd` method
+    * Add `updateCutoffs` function to permute FDR for proportionality
+        * Add `@metric`, `@ivar` and `@alpha` slots to calculate metric exactly
+        * Count rho > cutoff and cor > cutoff as positive results
+        * Count phi < cutoff and phs < cutoff as positive results
+* Update `propd` object backend and API
+    * Heavily revise documentation to harmonize `propd` with `propr`
+    * Fix bug where zeros still get replaced for `lrm` calculation
+    * The `updateCutoffs` function is now an S4 method
+    * Turn `@theta` slot into `@results` slot
+    * Rebuild `pd.d` and `pd.e` objects
+* Update `aldex2propr` method
+    * Update `aldex2propr` and `[` to disable `@results` and `@permutes`
+    * Add `lr2glm` and `aldex.glm` functions
+* Add methods
+    * Add `qtheta` to calculate a cutoff of theta for a given p-value
+* Remove methods
+    * Remove `differentialCheck`
+    * Remove `prop2prob`
+    * Remove `abstract`
+    * Remove `initialize`
+    * Remove `adjacent`
+* Revise vignettes
+
 ## propr 3.5.1
 ---------------------
 * Update CITATION file and README
@@ -166,7 +204,7 @@
 ---------------------
 * Modified `propd` Class
     * Added `@weighted` slot now used by `updateCutoffs`
-* Implement # Initialize @theta for "weighted theta" calculation
+* Implement @theta for "weighted theta" calculation
     * Added `wtmRcpp` for weighted mean calculation
     * Added `wtvRcpp` for weighted variance calculation
     * Added `lrm` function with optional weighted calculations